backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
requests
pandas
openpyxl
google-generativeai
